ml-service/requirements.txt
# ML Service - Python requirements
# Data Science & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
# tensorflow>=2.13.0  # Not compatible with Python 3.14 yet
# torch>=2.0.0  # Optional for future use
# prophet>=1.1.4  # Optional for future use

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Database
# psycopg2-binary>=2.9.9  # Not needed for SQLite
sqlalchemy>=2.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Data Processing
scipy>=1.11.0
# statsmodels>=0.14.0  # Optional for advanced time series

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Date/Time
python-dateutil>=2.8.2

# Logging & Monitoring
loguru>=0.7.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Model Management
joblib>=1.3.2
# pickle5>=0.0.11  # Not needed for Python 3.8+
